dnf-json: Test dnf-json to make sure it updates the cache correctly
This adds a helper, fake-rpm, which uses rpmfluff to create fake rpm
files for testing.

The test creates a repo, adds a fake rpm to it, depsolves the first
package version and then sets the atime and ctime of the dnf cache to 72
hours in the past.  It then adds a new version of the fake package,
updating the repodata and the timestamps. If it fails if the 2nd depsolve
doesn't return the new package version.